% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_vcf.R
\name{tidy_vcf}
\alias{tidy_vcf}
\title{tidy_vcf}
\usage{
tidy_vcf(
  raw_vcf = NULL,
  clean_geno = NULL,
  col_variant = "var_id",
  check_dim = TRUE,
  output_directory = NULL,
  output_name = "tidy.vcf",
  bin_path = list(bcftools = NULL, tabix = NULL),
  nb_cores = 1
)
}
\arguments{
\item{raw_vcf}{A \code{character}. The path to the raw merged VCF (format .vcf or .vcf.gz).
Default is \code{NULL}.}

\item{clean_geno}{A \code{character}. The path to the corrected genotype matrix (output of \code{check_genotype()})
with variants in rows and samples in column. The first column should be the unique ID of variants,
which is composed of variant's chromosome, position, reference allele and alternative allele,
separated by "_". Default is \code{NULL}.}

\item{col_variant}{A \code{character}. The column name of variant ID in \code{clean_geno}. Default is \code{var_id}.}

\item{check_dim}{A \code{logical}. Does the number of samples and variants should be checked in data from
\code{raw_vcf} and \code{clean_geno}. Default is \code{TRUE}.}

\item{output_directory}{A \code{character}. The path to the output directory. Default is \code{NULL}.}

\item{output_name}{A \code{character}. The output VCF names. Default is \code{tidy.vcf}.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of \code{bcftools} and \code{tabix}.
Final VCF will be ordered and compressed in gz format if \code{bcftools} is available, and will be
indexed if \code{tabix} is available. Default is \code{NULL} for all binary paths.}

\item{nb_cores}{A \code{numeric}. The number of CPUs to use. Default is \code{1}.}
}
\description{
Correct missing genotype in VCF using corrected genotype matrix.
}
