% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_coverage.R
\name{compress_coverage}
\alias{compress_coverage}
\title{compress_coverage}
\usage{
compress_coverage(
  sample_sheet,
  output_directory = NULL,
  chromosomes = NULL,
  min_depth = 8,
  bin_path = list(awk = "/usr/bin/awk", bgzip = "/usr/bin/bgzip", bgzip2 = NULL),
  nb_cores = 1
)
}
\arguments{
\item{sample_sheet}{A \code{data.frame}. A data frame containing samples \code{id} and
the full path to coverage files \code{cov_file}.}

\item{output_directory}{A \code{character}. The path to the output directory.}

\item{chromosomes}{A vector of \code{character}. Chromosomes to keep.
(same nomenclature as chromosome name in coverage file).
If argument not specified (default is \code{NULL}), this filter will be turned off.}

\item{min_depth}{An \code{integer}. The minimum depth to keep. Default is \code{8}.}

\item{bin_path}{A \code{list(character)}. A list giving the binary path of \code{awk} and \code{bgzip} (or \code{bgzip2})
in case of compressed coverage file.}

\item{nb_cores}{An \code{integer}. The number of CPUs to use. Default is \code{1}.}
}
\description{
Compresses coverage file (output of \verb{samtools depth}) into contiguous segments based on position.
This function works with whole genome or any specified chromosomes.
Each sample is supposed to have one coverage file.
The coverage file should not have header and contain 3 mandatory columns for chromosome, position
and depth, with no header.
Attention, currently only supports coverage file in raw text or .gz/.bz2 compressed format.
}
